refactor: implement built-in AI core functionality and associated tests
