+++ src/shared/built-in-ai/core/built-in-ai.test.ts
