+++ src/shared/built-in-ai/core/useBuiltInAI.ts
