+++ src/shared/built-in-ai/useRewriter.ts
@@ -1,5 +1,5 @@
-import { type CreateOptions } from "./built-in-ai";
-import { type UseBuiltInAIResult, useBuiltInAI } from "./useBuiltInAI";
+import { type CreateOptions } from "./core/built-in-ai";
+import { type UseBuiltInAIResult, useBuiltInAI } from "./core/useBuiltInAI";
 
 export const useRewriter = (
   options?: CreateOptions<"Rewriter">,
